Removing images, update setup script, and format python script
